Fix #40
Add new parameter --params that shows the entirety of the
parameters of the executable. Both the executable and the
parameters originate from the cmdline field of the process,
split by the first null character.

The parameters are currently chunked into a fix amount of
characters (at the time: 40, this can change) and then joined
by newlines.