Fix #36
When no -d/--directory is provided to the cloud_scanner main
script, a temporary directory under /tmp/cloud_scanner is used to
hold the files.

This directory and its contents will be deleted after the script
is completed. If caching/persistence is desirable, an already
existing directory should be provided with -d/--directory.