Fix #45
Preliminary support for socket ports.

Among the open files of a process that were displayed using
`--fd` or `--onlyfd`, it used to be the case that the type
(e.g. socket, pipe, etc) together with the inode was printed.

This commit replaces the inode number with the port number as
inodes numbers are not very interesting. Some port numbers could
not be found, this has to be further investigated.