Fix #78
The `--filesystem` view not trims the displayed information at the
end of the line depending on the terminal width, like the ps tool
does. This makes the output more readable.

However, when the output is not a tty (e.g. when using less or
grep), the output is fully written as normal. This may create very
long lines that are not pretty but have complete information.