Fixed bug when process tree was not shown.
This update fixes a crashs when attempting to show the process tree. The problem was that the viewer was not able to find the children of a process. The viewer is now slightly refactored to allow a function for getting the children instead of having a map for those. Also, the DAW now provides a function to find the children for a process.