Allowed filtering of user and group when outputting the running processes.
This commit adds the ability for the process tree to honor the user/gid filters
(i.e. --user, -g, ...). When the filters are applied, the processes will not be
printed as trees. This also adds two new helpers to the ProcessData wrapper to
check whether a process has a specific uid or gid.

This implements #88.