The scanner can now show threads.
The scanner is now able to display threads. For this, probe.py was extended to collect the same command line information on threads as on processes. Also, when the proc output tree is built, a function will be called each time to check for data variations from the thread to their parent. If something is diffrent, the thread will be displayed with the differences highlighted.

This implements #109