run_tests: support configurable threads count, reduce default to one
It turned out that the parallelization is actually reducing throughput,
probably due to cache effects or overload on the block device I/O.

This change switches from manually setup threads to a thread pool based
solution. Also some more information about the kind of tests being run
and a lock for serialization of stdout output.