global: rename occurrences of "security scanner" to Hamster
Since "cloud scanner" and later "security scanner" have been way too
generic names for a way too long time this is finally the decision to
use a more unique and rememberable name.

The idea behind it is that the scanner collects ("eats") data and stuffs
it away just like a Hamster does. And then uses it at the right time
again ;-)